bugfix: update packnames after identify
if you have two (or more) items of a kind in your pack
and only identify one of it the other one is not updated.
this patch goes through the list and updates all items which
have the same type and subtype to update its name also.
you could do the same by dropping and regetting the item
manually.